Massive input test created
