+++ output/actual/levels7.txt
@@ -1,0 +1,38 @@
+1027 6.00
+1020 3.00
+1012 2.73
+1014 2.67
+1025 2.57
+1008 2.00
+1023 2.00
+1024 2.00
+1028 1.20
+1029 1.00
+1035 1.00
+1003 0.86
+1038 0.86
+1040 0.86
+1016 0.82
+1018 0.67
+1034 0.67
+1036 0.40
+1009 0.29
+1032 0.18
+1031 0.00
+1033 -0.29
+1019 -0.35
+1022 -0.50
+1039 -0.67
+1011 -0.89
+1017 -0.95
+1005 -1.00
+1002 -1.25
+1001 -1.37
+1013 -1.65
+1007 -1.80
+1021 -2.00
+1006 -2.40
+1030 -3.00
+1010 -3.23
+1004 -4.00
+1037 -4.50
