+++ output/actual/tournaments6.txt
@@ -1,0 +1,30 @@
+1009
+999
+582.79
+Germany
+63
+36
+1019
+999
+605.08
+Olman
+52
+35
+1015
+980
+479.51
+Students house
+57
+36
+1001
+964
+579.44
+Makefile is overrated
+54
+35
+1015
+995
+550.19
+Greece
+53
+35
