+++ output/expected/levels6.txt
@@ -1,0 +1,40 @@
+1027 6.00
+1020 2.25
+1014 2.00
+1029 2.00
+1034 2.00
+1008 1.11
+1025 1.00
+1012 0.75
+1018 0.75
+1003 0.59
+1019 0.43
+1028 0.40
+1009 0.36
+1026 0.29
+1023 0.22
+1031 0.00
+1001 -0.09
+1015 -0.09
+1016 -0.20
+1022 -0.22
+1040 -0.22
+1033 -0.40
+1017 -0.43
+1032 -0.46
+1024 -1.00
+1002 -1.24
+1011 -1.50
+1006 -2.00
+1021 -2.00
+1030 -2.00
+1035 -2.00
+1038 -2.00
+1013 -2.15
+1007 -2.33
+1036 -2.57
+1005 -2.64
+1004 -4.00
+1039 -4.40
+1037 -4.50
+1010 -5.53
