+++ output/expected/tournaments3.txt
@@ -1,0 +1,6 @@
+1010
+6969
+2197.30
+Mother russia
+10
+9
